[CoreBundle] Api clean some methods. (#3065)
* [CoreBundle] Api workspace clean some methods.

* [CoreBundle] Some API improvements.

* More api fixes

* Removing tests from removed methods.

+++ main/core/API/Crud.php
@@ -18,7 +18,6 @@ class Crud
 
     const COLLECTION_ADD = 'add';
     const COLLECTION_REMOVE = 'remove';
-    const COLLECTION_REPLACE = 'set';
 
     /** @var ObjectManager */
     private $om;
@@ -79,9 +78,12 @@ class Crud
         // creates the entity if allowed
         $this->checkPermission('CREATE', $object, [], true);
 
-        $this->dispatcher->dispatch('crud_pre_create_object', 'Crud', [$object]);
-        $this->om->save($object);
-        $this->dispatcher->dispatch('crud_post_create_object', 'Crud', [$object]);
+        $event = $this->dispatcher->dispatch('crud_pre_create_object', 'Crud', [$object]);
+
+        if ($event->isAllowed()) {
+            $this->om->save($object);
+            $this->dispatcher->dispatch('crud_post_create_object', 'Crud', [$object]);
+        }
 
         return $object;
     }
@@ -105,10 +107,12 @@ class Crud
 
         // updates the entity if allowed
         $this->checkPermission('EDIT', $object, [], true);
+        $event = $this->dispatcher->dispatch('crud_pre_update_object', 'Crud', [$object]);
 
-        $this->dispatcher->dispatch('crud_pre_update_object', 'Crud', [$object]);
-        $this->om->save($object);
-        $this->dispatcher->dispatch('crud_post_update_object', 'Crud', [$object]);
+        if ($event->isAllowed()) {
+            $this->om->save($object);
+            $this->dispatcher->dispatch('crud_post_update_object', 'Crud', [$object]);
+        }
 
         return $object;
     }
@@ -124,10 +128,13 @@ class Crud
     {
         $this->checkPermission('DELETE', $object, [], true);
 
-        $this->dispatcher->dispatch('crud_pre_delete_object', 'Crud', [$object]);
-        $this->om->remove($object);
-        $this->om->flush();
-        $this->dispatcher->dispatch('crud_post_delete_object', 'Crud', [$object]);
+        $event = $this->dispatcher->dispatch('crud_pre_delete_object', 'Crud', [$object]);
+
+        if ($event->isAllowed()) {
+            $this->om->remove($object);
+            $this->om->flush();
+            $this->dispatcher->dispatch('crud_post_delete_object', 'Crud', [$object]);
+        }
     }
 
     /**
@@ -140,25 +147,27 @@ class Crud
     public function deleteBulk($class, array $data, array $options = [])
     {
         $this->om->startFlushSuite();
+
         foreach ($data as $el) {
             //get the element
             $this->delete($el, $class, $options);
         }
+
         $this->om->endFlushSuite();
     }
 
     /**
-     * Patches a collection in `object` allowing to add or remove elements in it.
+     * Patches a collection in `object`. It will also work for collection with the add/delete method.
      *
      * @param object $object   - the entity to update
      * @param string $property - the name of the property which holds the collection
-     * @param string $action   - the action to execute on the collection (aka. add/remove)
-     * @param array  $elements - the list of elements on which to execute `action`
+     * @param string $action   - the action to execute on the collection (aka. add/remove/set)
+     * @param mixed  $elements - the collection to patch
      */
     public function patch($object, $property, $action, array $elements)
     {
-        // retrieves correct method to call on entity
-        $methodName = $action.substr(ucfirst(strtolower($property)), 0, -1);
+        $methodName = $action.ucfirst(strtolower($property));
+
         if (!method_exists($object, $methodName)) {
             throw new \LogicException(
                 sprintf('You have requested a non implemented action %s on %s', $action, get_class($object))
@@ -167,12 +176,39 @@ class Crud
 
         //add the options to pass on here
         $this->checkPermission('PATCH', $object, [], true);
+        //we'll need to pass the $action and $data here aswell later
         $this->dispatcher->dispatch('crud_pre_patch_object', 'Crud', [$object]);
-        $methodName = $action.substr(ucfirst(strtolower($property)), 0, -1);
 
         foreach ($elements as $element) {
             $object->$methodName($element);
         }
+
+        $this->om->save($object);
+        $this->dispatcher->dispatch('crud_post_patch_object', 'Crud', [$object]);
+    }
+
+    /**
+     * Patches a property in `object`.
+     *
+     * @param object $object - the entity to update
+     * @param string $action - the action to execute on the collection (aka. add/remove/set)
+     * @param mixed  $data   - the datas that must be set
+     */
+    public function replace($object, $property, $data)
+    {
+        $methodName = 'set'.ucfirst(strtolower($property));
+
+        if (!method_exists($object, $methodName)) {
+            throw new \LogicException(
+                sprintf('You have requested a non implemented action \'set\' on %s', get_class($object))
+            );
+        }
+
+        //add the options to pass on here
+        $this->checkPermission('PATCH', $object, [], true);
+        //we'll need to pass the $action and $data here aswell later
+        $this->dispatcher->dispatch('crud_pre_patch_object', 'Crud', [$object]);
+        $object->$methodName($data);
 
         $this->om->save($object);
         $this->dispatcher->dispatch('crud_post_patch_object', 'Crud', [$object]);
